feat: replace the export app modal with one that uses trailhand components
